:zap: Drupal JS deps should be first-party requests
The version of jQuery is also changing to match the latest
supported version of jquery_update module which the Drupal
themes which use CD Header/Footer have to use. Additionally
the copy of jQuery is using the same noConflict mode that
appears by default on Drupal sites, requiring jQuery code
to be wrapped in a closure to expose the $ shortcut.